AI Update / Objects
Implemented A* tracking and applied it to red ghost. ghost now appropriately follows player. Pickup method created for objects in map, when pacman crosses the object is removed and a score is added. Large pellets cause red ghost to scatter to home square and when caught by pacman its position is reset and the score is updated.